pyproject.toml
[project]
name = "ml-project"
version = "0.1.0"
description = "機械学習のための新モデルの開発・構築・デプロイ"
requires-python = ">=3.11"
dependencies = [
    "torch",
    "numpy",
    "matplotlib",
    "ultralytics>=8.4.12",
    "transformers>=5.1.0",
    "sentencepiece>=0.2.1",
    "pillow>=12.1.0",
    "faiss-cpu>=1.13.2",
    "opencv-python>=4.13.0.92",
    "timm>=1.0.24",
]
